Move MPG123 class as a plugin
