+++ src/filters.py
@@ -1,44 +1,9 @@
-# ALLOWED FILTERS : Available for patches, meaning, allow only for instance
+# # ALLOWED FILTERS : Available for patches, meaning, allow only for instance
 q49=ChannelFilter(1)    # Filter by hardware / channel
 pk5=ChannelFilter(2)    # Filter by hardware & channel
-fcb=ChannelFilter(9)
-hd500=ChannelFilter(9)
-gt10b=ChannelFilter(16)
-
-# By name
-ch1=ChannelFilter(1)
-ch2=ChannelFilter(2)
-ch3=ChannelFilter(3)
-ch4=ChannelFilter(4)
-ch5=ChannelFilter(5)
-ch6=ChannelFilter(6)
-ch7=ChannelFilter(7)
-ch8=ChannelFilter(8)
-ch9=ChannelFilter(9)
-ch10=ChannelFilter(10)
-ch11=ChannelFilter(11)
-ch12=ChannelFilter(12)
-ch13=ChannelFilter(13)
-ch14=ChannelFilter(14)
-ch15=ChannelFilter(15)
-ch16=ChannelFilter(16)
-
-noch1=~ChannelFilter(1)
-noch2=~ChannelFilter(2)
-noch3=~ChannelFilter(3)
-noch4=~ChannelFilter(4)
-noch5=~ChannelFilter(5)
-noch6=~ChannelFilter(6)
-noch7=~ChannelFilter(7)
-noch8=~ChannelFilter(8)
-noch9=~ChannelFilter(9)
-noch10=~ChannelFilter(10)
-noch11=~ChannelFilter(11)
-noch12=~ChannelFilter(12)
-noch13=~ChannelFilter(13)
-noch14=~ChannelFilter(14)
-noch15=~ChannelFilter(15)
-noch16=~ChannelFilter(16)
+# fcb=ChannelFilter(9)
+# hd500=ChannelFilter(9)
+# # gt10b=ChannelFilter(16)
 
 # Control Filter Channel (cf)
 cf=~ChannelFilter(9)
