+++ src/start.sh
@@ -12,7 +12,7 @@ sudo mkdir -m777 -p $flash
 count=$(df | grep -c sda1)
 if [ $count -eq 0 ]
 then
-	sudo mount /dev/sda1 $flash
+	sudo mount /dev/sdb1 $flash
 fi
 
 ln -sf $audio /tmp/soundlib
